session to parameter
After authentication session was being started. It works in case of
browser but not from app. So to make it easy while development and to
save time, after successful signin I am sending a parameter
'authenticated' 1 if user has been signed in using google else 0. I know
it's not secure but that'll be handled later.